Small fixes for using remote memory operations
Small errors in the use of REM_STX -- you need to be careful about how
you specify the arguments, as the macro does casting and doesn't do type
checking.

Additionally, it appears there are several bits that may be set by acquire
that are not part of either the status or the address: strip these out using
the GET_ADDRESS macro.

git-svn-id: https://uhpc.nbgisp.com/svn/runnemede/trunk@2314 b8300b04-83cd-4eb7-bcdf-78d2b5523b00