Initial check in of libocr-glibc.
This is a shim between glibc and ocr.  Included are changes to newlib
sys/ocr sources to allow for multi-use; i.e. xstg newlib, x86 C newlib,
and x86 C and C++ glibc support. Using weak_alias instead of #defines
to provide glibc compatible syscall names.

This implementation requires that the user application use mainEdt()
instead of main() as the starting point.  In this case mainEdt() has
the same return type and arguments as a normal main().

Change-Id: I27801d42ccf2bcccb2963e075b6e74d31535222c