Added a simple commit hook to list bugs in a commit
This script is meant to help people add bug references
to their commit messages. It should catch most of them

It considers a bug to be fixed if a line containing a reference
to a bug is removed.

It considers this to be related to a bug if a reference to a bug
is added

Change-Id: I16fac656399d99bcb446cc74d494998068e49ba0