Adding 2 ROSE translation tools: threadTool and globals2struct.
threadTool adds __thread before each global and file static variable.

globals2struct finds each global and file static in an application.
It creates a data structure, where copies of the global and file
static variables will reside.  The tool initializes the values that
were originally initialized.  It then updates the reference to the
globals and file statics in the application.

Note that no OCR support is emitted by these tools.

Change-Id: I835a04a85c73972045b22dfab68f002dd41c5d0b