Enhancements to MPI-lite to support the globals2db ROSE tool.
This checkin adds the functions and data structures needed
to support the code modications that the globals2db ROSE tool
will emit.  It provides the support needed to access and initialize
the new data block.

The globals2db ROSE tool is still a WIP.

These modifications were tested on the MPI-lite version of Stencil1D.

Change-Id: Ib4612b046eea09b9d147304a0429d7525a319fa6