globals2db - a ROSE tool which creates an OCR DB for globals.
This ROSE tool finds all global and file static variables in an
application.  It creates an OCR Data Block, and places the globals
in it.  It then updates the references in the application.

Three tests are included in this commit.  Two of them are negative
tests, and are cleverly named "errorX".

Change-Id: Icd746b6022145a12f5cee4bf0d6826a6c3ef4435