Add Labeling to mpilite - turned off
Use of labeling will be turned on if mpi_ocr.h: EVENT_ARRAY==0. For now
it is 1 because you need a different config to use labeling, and everything
would break. So MPI-Lite behavior is not changed with this checkin.

But, all of the code for labeling is here, and works for stencil1d (havent
tried others with new configs):
mpi_ocr.h: creates a 3-D map, passes it to Ranks, and is stored in
messageContext.

mpi_ocr_messaging.c: the only file actually using labeling in ocrEventCreate
and ocrLegacyBlockProgress

mpilite.c: New #if ! EVENT_ARRAY versions of send/recv, isend/irecv, which
call the new fns in mpi_ocr_messaging.c which use labeling.
There is also Aggressive Non-blocking for i{send/recv}, and wait/test. They
can use the new mpiOcrTry{Send,Recv} in mpi_ocr_messaging.c; letting them
get started earlier. But down side is, messages might get out of order
depending on usage pattern. So this can be turned on with a command line
flag (the first): "-a" for experimenting. Only works with labeling.

makefile.x86: build mpi_ocr_messaging.o

stencil*: more examples, but not the default version use "make TARGET=<new file> run" to try them

Change-Id: If4c081b231ff41f56cb67784cfa590156dd59515