Add CONFIG_FLAGS to config_generate step
Only for normal (not visualizer) build of make-regular.inc.
Anyone who needs LABELED guid provider in config can give this
variable a value, like "--guid LABELED" in their makefile, e.g.,
for MPILite, or David Scott's usage (and Chandra's plans to do so).
This change is required before use of labeling can be turned on in
MPILite.

Change-Id: Ie2299bafd0b1b7d31fc023034d2fe0b080556e07