Added support for function statics; consolidated 2 ROSE tools.
This checkin adds support for function statics.  Function statics
will also be added to the OCR DB, and initialized (when applicable)
in the function.

This checkin also consolidates the source code for globals2db and
globals2struct under the directory name globals2db.  Both tools
are built when the Makefile is invoked.  Tests for both tools have
been moved under the globals2db directory as well.

The older source code and tests will be removed in a subsequent
check-in.

Change-Id: I039f3854b3e45e476fa7755c800aeba8591f7352