app-lite makefiles now make LABELED configs
Changed the 6 apps' Makefile.x86 to have
    CONFIG_FLAGS := --guid LABELED
which will cause new config generator to make LABELED guid generator.
This is necessary before MPI-Lite's use of labeled guids can be turned on.
Also in a few makefiles, removed the "-r 4" restriction of the default
workload to only use 4 ranks. Now it will use as many ranks as there are
workers.

Change-Id: Idf1fb8ae94837fbc89dec4fa2e77a1401cab274d