Added support for processing system-defined and user-defined types.
System-defined and user-defined types must be entered in the DB
structure the same  way that they are originally present in the
source code.   This required capturing the original string, and
using it when creating the DB structure.

This implementation relies on the user to put all user-defined
classes and structs and typedefs used by the globals and the file
and function statics in an include fle.  The inclusion of the
include file created by this tool, "ffwd.h", should be placed
after all of the other includes.  It has to be done this way for
the user-defined types will be known to the compiler when processing
the structure that will be used by the Data Block.

Also added support for checking the symbol tables with lesser scopes
when the global scope symbol table does not contain the symbol
of interest.

Did some clean-up work as well.

Change-Id: I77db421e5a3fd61f8f732816b6ed4266acea2e22