Update to library Makefiles to make x86 target consistent.
This change makes the x86 target in library Makefiles to indicate
that the x86_64 targeted clang/llvm tool chain is being used; this
was originally called a x86_clang target in the some of the
Makefiles.  The original x86 target was being used to indicate
that a native gcc was being used; this Makefile target has now been
renamed x86_native.

This change affects newlib, ocrscaffold, ocr-glibc shim, and the
apps/legacy examples and is reflected in their associated README
files.

Change-Id: If87ae133db5c0c1e15d658bdcec788a524079669