Updated globals2db, Stencil1d-lite and CoMD-lite.
Incorporated code review comments.

Changed the way that CoMD-lite calls globals2db.
Originally this support was in the Makefile.  However,
That did not work well with Jenkins, which calls the
platform-specific Makefile directly.  Now this support
is in the platform specific Makefile.  Using globals2db
cannot be done from Jenkins yet because the hll/rose
directory is not copied to testDir, and so the build fails.

Added scripts to Stencil1D-lite to use 8 and 16 threads.

A little bit of cleanup in globals2db.

Added License stuff to defuse test.

Change-Id: I78141ae8e2c0e6b42e68406671d1c7c7a35fa448