Disabled labeled GUIDs in MPI-Lite.
Temporarily disabled labeled GUIDs in MPI-Lite for the Apps Workshop.
Labeled GUIDs are not scaling, and they do not work for 8 or 16 threads
in Stencil1Dlite.  This will be re-enabled after the Workshop.

Add support to Jenkin to copy over the ROSE tool sources.

Updated XSBench to use globals2db.  Updated other *lite Makefiles.

Fixed a bug in globals2db which emitted the function static
initialization more than once.

Change-Id: I3ef84dda0a6d7720872697b428976861e0d93b60