Have a working object!
Tthe 'Legacy C' code (mainEdt) now calls on
newlib to perform it's operations.  The cmain (main()), which does
all the ocr initialization calls on the libocr to perform it's
operations.
Currently, the first printf in mainEdt is not working; it fails in
NL___sfvwrite_r, where it calls on 'cantwrite' which returns true,
and therefore aborts.

Change-Id: I910349643fed44a8e812204c75310f187fab4d1e