mpilite on x86-mpi: guess max ranks
mpi_ocr.c: add maxWorkers() to guess the number of workers
available to use for ranks. Use ocrAffinityCount to figure out if
there is more than 1 policy domain (i.e., in x86-mpi). If so,
1 thread in each PD is used for MPI. ocrNbWorkers() tells how many
workers are in current PD, so subtract by 1 and mul by #PDs. And
subtract a fudge factor. If only 1 PD, assume non-distributed and all
threads are available for ranks.
Need to include affinity.h, and

mpi_ocr.h: turn on the affinity extension in mpi_ocr.h.

Change-Id: I23c87c38db06d50d8688421db5c7a0fdd4c5810c