Fixed a bug in globals2db which caused CoMD-lite to fail.
CoMD-lite was failing to build after being translated
by globals2db.  The include file rose_ffwd.h, which is
generated by globals2db, was not included in rose_CoMD.c.
It was included in other translated files, but not in
this file.

This fix uses a different version of the Sage function
insertHeader(), which requires the source file.

Tested with CoMD-lite, XSBench-lite and SNAP-lite.

Change-Id: I90b37f6cbe2df057ccf634f8030f5ad98dff5c81