Updated miniAMR so globals2db can digest it.
Added the globals2db support to Makefile.x86.
Added guards to the include files so they will only be included once.
Expanded param.h in main.c.  Param.h contains the initilialization
of many globals.  When globals are initialized in .c files, the
reference is updated to use the DB pointer.  Today, globals2db
does not emit updated include files.  Expanded param.h in main.c
so the initialization will occur properly.

There is probably a way to emit updated include files in ROSE.
But there is an easy work-around to this issue, and the use
of globals2db is quite limited.

Change-Id: Id3e7b9650cc10a563dda5f7cdd4844c770154ff0