miniAMR-lite now works with globals2db.
As you know, globals2db takes the globals in an application
and places them in an OCR DB. It then replaces all references
to the global with a DB pointer.

miniAMR used globals as its dimensions when declaring arrays.
ROSE does not support replacing the indices of an array in
the declaration.  I added local variables to miniAMR which
were set to the value of the global, and used the local
variables in the declaration.

miniAMR also relied on the compiler to do initialization of
global variables and malloc'd data.  I added initialization
where required.

globals2db creates a tool generated variable to guard the
function static initialization.  This TGV was not initialized,
and that prevented the function static from being initialized.
I added this support.

Tested on x86 and x86-mpi.

Change-Id: I08b115a53d3f2cc1eb218b0214dcf5d0d0b81736