Updates to globals2db - initialization support.
Support has been added to globals2db to initialize global,
file static and function static variables that were not
initialized in the application.  The application depended
on the compiler to do this.  globals2db now handles this.

Support has been added to handle initialization of global
variables that were declared in header files.

Also, an error was discovered in the placement of the
initialization of function statics when the function static
was declared in the middle of the function.  The initialization
was done too soon, and the value in the datablock was not
initialized with the correct value.  This has been fixed.

A new test case has been added to globals2db.
Initialization code that I added to miniAMR has been removed.

Change-Id: Ifdb68601d24c6be51f44f7521afc44b10bc1c2a2