OCR vsn miniAMR: totally unrefined stencil works. Copyrights updated.
In the ...refactored/continuationcloning/intel folder, this check-in merely
updates the copyright notice to reflect Intel contributions.

In the ...refactored/ocr/intel folder, this check-in does that, but it also
checks in all support so far for accomplishing successful running of
miniAMR, subject to the limitation that NO REFINEMENT is performed (yet).

(Makefile.x86 presently specifies a 6*7*8 mesh, so increase dependency slots
to at least 6 sets of 64, in common.mk.  Don't forget to uncomment line.)

Change-Id: Ibd2bd3c93889ecc525bbc75e230da480e6e71973