Updates to MPI-Lite Unit Testing.
This commit contains the functionality for unit testing.
The top-level Makefile invokes the local Makefile in each
listed subdirectory.  Both x86 and x86-mpi are tested.
TG will be added later.

The tests are self-verifying.  The passing criteria are
tested in the local Makefile, and a PASSED/FAILED message
is printed to a log file.  The number of passing/failing
tests is printed to the screen.

Directions for running the test suite ("make") and adding
new tests are in the README.

Change-Id: I32b8216743c5b2a8d42742b181a1764333db04e6