global2db: allow implicit dimension [].
Added the support for the implicit size dimension []
when declaring a global or file static array or matrix.
The implicitly defined global variable must be
initialized at declaration (not from a file).  This
restriction exists because the globals are placed in
a structure which is contained in a data block.
globals2db must allocate the correct amount of memory
for the data block.

Also fixed a string manipulation bug in globals2db.

Updated some of the globals2db tests.

Change-Id: If4b53395c1f92333439e0ce6d452b87f83cd68b1