miniAMR now working with MPI_Comm_split().
Tested the new communicator support with miniAMR.
miniAMR is the only proxy app that uses communication
groups.  It calls MPI_Comm_split() on MPI_COMM_WORLD
so only a subset of the world communicates with
each other.

Fixed a couple of issues in the MPI-Lite communication
support that was not exercised by the kernel tests.

Change-Id: I185e2393d538905f2fc670e631cac47f21975206