Update futures API
A "future" is an EDT with an output event. Since the output events are
once-type (or latch-type), this makes it difficult to correctly set
dependences on the output event (avoiding races). To avoid this problem,
we delay running "futures" until the user explicitly calls Ready() on
the DelayedFuture object returned by template.CreateFuture().