Remove CreateFuture method from public API
It doesn't make any sense to supply all the dependences to a future
because you can't use the output event if you do that, which implies
that you should always use CreateFuturePartial instead.

Also optimized for empty depv (using nullptr instead of creating
an entire vector of UNINITIALIZED_GUID values).