Graph500 modified to work on TG (with OCR)
The Graph500 source code was moved from .cpp to .c and the
source code was modified to make the features not currently
available on TG (like files) optional. Furthermore, the
functions used by the program (like random number generator,
assert or fflush) that are not available on TG were either
reimplemented or replaced with empty implementations.

Change-Id: I64197820a39ec9a013fd5f450ba0db18bde10410