autogen IF_THEN_ELSE pattern and recursion
RedMine ticket # 966
The IF-THEN-ELSE pattern was added to autogen graph description.
This makes explicit the use of a task for switching purposes.
Recursion uses this IF-THEN-ELSE pattern and the newly implemented
concept of shared connection.
Consider the graph edge A->C and B->C, where A,B,C are ETDs.
Saying that B->C share a connection with A->C means that all events
handled by B->C are going to be the same as those for A->C, with the
exact same slot assignments.

Two black-box tests has been implemented, one for helloWorld, and one
for tailRecursion which examplifies the use of the IF-THEN-ELSE
pattern and shared connection. Please refer to their respective README.

Amended to remove build, install and tools executables and object files.

[Chandra]: Removed .pyc and fsim config files

Change-Id: Ied1f337bd5b29f5fd17ac9417e27108645c8148f