Jenkins: Adjusting the number of points in X-axis
Allowing to adjust the number of points in X-axis of the performance graphs.
The default number of points is set as 99.
Tumber of points can be changed using OCR_APPS_PLOT_POINTS environment variable.
OCR_APPS_PLOT_POINTS should be a +ve integer value.
Tor -ve or zero value of OCR_APPS_PLOT_POINTS, the default number of points will be used.
The same change is done in ocr.

Change-Id: Ib48c3077ce419287da6d53bf8422ca5915666fae