Autogen binary fork-jouin tree and debug facility
Redmine # 977

A FOR-pattern was made explicit in Autogen.  It allows one to have
a one to many creation process in Autogen.
It was use to make a full binary-tree fork-join example.
WIth some re-coding fo the customer provided code, any N-ary tree
could be created.
A run-time operation tracing tool was added to Autogen in order to facilitate
debugging OCR operations.  See README fro more details.

Change-Id: Ia277384c4f74d06f95cc4c70437e4221f3054b81