Initial Nekbone code addition
The commit has 3 sections: The baseline, the skeleton and the refactored.
The baseline is what one can get off the site
    https://asc.llnl.gov/CORAL-benchmarks/#nekbone
It is version 2.3.4

The skeleton directory is a pure C++ (i.e. no Fortran) skeleton
of what is in the baseline.  That means all the communications,
and data blocks are there but nothing is done to fill up the
data blocks.

The refactored directory has 3 parts: nek_src, ocr_src and unit_tests.
The nek_src is the C translates from the baseline code.
The ocr_src is where the OCR specific code resides.
The unit_tests directory is a small unit test framework for insure
proper working of the components.

Change-Id: I37b6420fe3e4c5d4a2921235b1deaf45098d28d7