libcxxabi - fixed issue with pthread build
The library needs per-thread storage for exception info and
if pthreads are enabled in the build it assumes that pthread
APIs are there and working and/or thread local storage. Unfortunatly,
when linking without pthreads (stubbed out), use of pthread APIs
caused a segfault. Since this is for XEs the solution is to put
the exception info in the XE scratchpad (module static data) which
works as long as we only ever do 1 thread per XE.
Also added changes to the C++ library makefiles to be able to enable
a debug build (-g).

Change-Id: If16de1b3589d4040ef391f622d23988f8644ba46