apps libcxxabi - changed XSTG to not need TLS
Currently, in a non-pthreads link, TLS is not available (no crt support)
So, for XSTG the thread-specific data was made just static local which
gets put in XE local memory. This is effectively thread specific since
we're only mapping a single thread to an XE. Also pthread support was
turned off for the x86 targets as we don't support that.

Change-Id: Ib1bdb6ed5e37fa033e940c4e1087af1e281902b1