Various changes for openmp, newlib, and libswtest.
Added support for initializing tls in newlib. Also added _ftbss
and _etbss symbols to the swtest linker script to be able to zero
the thread bss section. Prevented deadlock when trying to create
more than 7 child threads. In libomp cast fake varargs functions
to ones with explicit args.

The function __kmp_invoke_microtask takes a pointer to a vararg
function but gets passed functions with explicit arguments. It
requires casts to change the vararg functions to ones with
explicit arguments.

Reserve one XE in OpenMP for the monitor thread. Disabled
malfunctioning routine in OpenMP for getting gtid. Instead of
using stacks to determine the gtid, this patch forces OpenMP to
use the fallback method of pthread specifics.

Change-Id: I040a0824142db380aad2fd0b103c629fbb2a98c3