Various bug fixes and test improvements for openmp.
Fixed flock malfunctions due to mutex support not being linked in.
Added run-fsim-swtest-openmp-tests.sh to tools/regression_tests.
Changed the default number of test repeats from 20 to 2.

Removed errors which occurred when trying to make check in libomp
with a -j flag. Updated the legacy config files to prevent stack
problems with omp. Increased the stack in ccfg.cfg to be at least
16KB, and placed the stacks in L2. Also updated omp to avoid race
conditions in setting the id variable.

Decreased the loop count for the openmp testsuite to 20. Previously
the loop count was 1000. This drastically reduces the time needed to
run some tests. Reduced the amount of memory allocated by openmp with
bget. Previously it allocated 1MB memory pools. Now it uses 8KB memory
pools.

Added pthread_malloc.c to legacy and regression_tests.

Change-Id: Ib69d93f3d5f54a97d93f663fd9ccc4d745457cdf