Reverted change to omp_atomic.c and added new rt-builtin support.
The omp_atomic test was modified earlier due to problem with clang's
injection of __sync_fetch_and_sub_4 function at the point of the
omp pragma "atomic".  This change was removed and the test is now
at it's original state.  To satisfy the link of __sync_fetch_and_sub
a new file, xstg_sync_fetch_and_sub.c was added to the
compiler-rt-builtins library.  This file contains versions of the
__sync_fetch_and_sub_1, *_2, *_4, and *_8 functions.  A strong_alias
was used to get around a clang compile error of "definition of
builtin function".

Change-Id: I3067ea613539f99204e1ec106e78e3e5e271d59f