Implemented timeout mechanism for fsim regression scripts
The timeout mechanism to selectively applies timeouts per script.
Currently there is only a timeout applied to the openmp script.
The timeout may be set/overridden by exporting TIMEOUT_SECONDS and
it may be disabled by setting it to 0.
The regression tests timeout was still running after tests completed.
Made sure that tests killed the timeout timer as they finished.

Change-Id: I71b3cdc5891f2ca266dcf49747871a26488f3a37