Optimized openmp threads for tgr and disabled manager thread.
The blocktime knob for openmp determines how long threads stay
active before sleeping and yeilding to other (non-openmp) threads.
This is unimportant for tgr because each XE is devoted to one
thread for the durration of that thread. So blocktime was set to
the maximum to keep threads from ever sleeping.

As a side effect of this, the manager thread is no longer needed,
so it is automatically disabled. Thus we have another free XE for
an openmp thread.

Change-Id: I8e6cb977e22765429ce309c318e824d08292bbd3